Revert "feat(runtimed): expose llm output text helpers"
This reverts commit c4bd93466e36df911e2ca5989f56359b0ada2988.
