+++ crates/runtimed-outputs/src/lib.rs
@@ -7,5 +7,4 @@
 #![cfg_attr(test, allow(clippy::unwrap_used, clippy::expect_used))]
 
 pub mod output_resolver;
-pub mod output_text;
 pub mod resolved_output;
