+++ packages/runtimed-node/src/session.cjs
@@ -103,10 +103,6 @@ class Session {
     return this._native.getCell(cellId);
   }
 
-  getCellOutputText(cellId) {
-    return this._native.getCellOutputText(cellId);
-  }
-
   createCell(source, options) {
     return this._native.createCell(source, options);
   }
