+++ python/runtimed/src/runtimed/_internals.pyi
@@ -594,7 +594,6 @@ class AsyncSession:
     def get_cell_source(self, cell_id: str) -> Coroutine[Any, Any, str | None]: ...
     def get_cell_type(self, cell_id: str) -> Coroutine[Any, Any, str | None]: ...
     def get_cell_outputs(self, cell_id: str) -> Coroutine[Any, Any, list[str] | None]: ...
-    def get_cell_output_text(self, cell_id: str) -> Coroutine[Any, Any, list[str] | None]: ...
     def get_cell_execution_count(self, cell_id: str) -> Coroutine[Any, Any, str | None]: ...
     def get_cell_ids(self) -> Coroutine[Any, Any, list[str]]: ...
     def get_cell_position(self, cell_id: str) -> Coroutine[Any, Any, str | None]: ...
@@ -706,7 +705,6 @@ class AsyncSession:
     def get_cell_execution_count_sync(self, cell_id: str) -> str | None: ...
     def get_cell_sync(self, cell_id: str) -> Cell: ...
     def get_cells_sync(self) -> list[Cell]: ...
-    def get_cell_output_text_sync(self, cell_id: str) -> list[str] | None: ...
     def get_cell_metadata_sync(self, cell_id: str) -> str | None: ...
     def get_runtime_state_sync(self) -> RuntimeState: ...
     def get_peers_sync(self) -> list[tuple[str, str]]: ...
